Allow base classes for system tests.
An abstract class, using Python's abc module, can be defined to
amalagate code from tests that require similar workflows. The check
for tests now uses an import rather than a regex check on the file.
Implemented a base class for the direct inelastic tests with MARI as a
first test.