Re #6169. Use new API export of MemoryStats.
I was thinking of using psutil (http://code.google.com/p/psutil/)
but then realised that this would require it to be installed on
any machine that was going to run system tests. Also, it doesn't
seem to work on the latest Mac OS's.
However, I was pleased to see that its report of the available
memory on Linux & Windows matches exactly that given by Mantid.