Updating EQSANSTransAPIv2 system test
**warning** this change must be reviewed from SNS people.

After removing the linear from CalculateTransmission, the result of
processing EQSANS(False) and EQSANS(True) does not differ now only 0.1,
it would be necessary to change the tolerance to 0.3. I did not know if
this new tolerance was acceptable, or, if it would be better to have
two answers.

BTW, the difference in the tolerance is related to the error estimation,
as described inside track for the ticket #6981