Re #6563. Don't get stuck waiting for license acceptance.
The dmg has the GPL license embedded within it, which the
user is required to accept before the volume is mounted.
This accepts it by piping acceptance into the command. The
output to /dev/null deals with the fact that the license is
multi-page (effectively it scrolls to the bottom which is where
the 'Y' needs to be entered).