Re #10684 Decreased the accuracy of mari reduction wrapper test
to ignore changes in background removal. ReductionWrapper test verifies internal validation and this validation is not intended to be fiddled with in a way, other tests currently are. By its nature, it should not, so let's do it for the time being until decided is it better to remove or left background for vanadium and modified references files accordingly